useDataApiWithPrefCodeHook() 用のファイルを作成し、移動
